+++ R/is_constant.R
@@ -1,20 +1,13 @@
 #' @rdname tests
 #' @export
 
-is_constant <- function( x, na.action=na.omit ) {
+is_constant <- function( x, na.rm=TRUE ) {
   
   if( all( is.na(x) ) ) return(NA)
-  if( ! is.numeric(x) ) stop( "monotonicity can only be determined for numeric vectors.")
   
-  if( ! is.null(na.action) ) x <- na.action(x)
+  if( na.rm )
+    x <- na.omit(x)
   
   all( x == x[1] )
   
 }
-
-
-#' #' @rdname tests
-#' #' @export 
-#' is.constant <- function(...) { 
-#'   warning( "'is.constant' will be removed. Use 'is_constant' instead.")  
-#' }