+++ R/is_incremental.R
@@ -1,36 +1,39 @@
-#' is_incremental 
-#' 
-#' Test if a field is incrementally increasing or decreasing
-#'
-#' @param x numeric 
-#' @param step integer; step size for increment
-#' @param na.action NULL or function to pre-process data
-#' 
+# is_incremental 
+# 
+# Test if a field is incrementally increasing or decreasing
+#
+# @param x numeric 
+#' @param step integer; step size for `is_incremental`. (Default: 1)` 
+# @param na.rm logical; whether to omit `NA` values. (Default: TRUE)
+# 
 #' @details 
 #' 
-#' `is_incremental` determines if x is incremental, i.e. ordered and equally 
-#' spaced.
+#' `is_incremental` determines if x is incremental, i.e. monotonic and equally 
+#' spaced. 
 #' 
 #' `is_uniform` is a wrapper around `is_incremental` with `step=NULL`
 #' 
 #' @return 
 #'   logical 
 #'   
-#' @seealso 
-#'  - [tests]
-#'     
 #' @examples 
 #' 
 #'   is_incremental(1:5 )
 #'   is_incremental( c(1,2,5))
-#'  
+#'   
+#'   is_incremental(1:5, step=NULL)
+#'   is_uniform(1:5)
+#'
+#' @rdname tests
 #' @md    
 #' @export 
 
-is_incremental <- function(x, step=1, na.action=NULL  ) { 
+is_incremental <- function(x, step=1, na.rm=TRUE  ) { 
 
   if( all( is.na(x) ) ) return(NA)
-  if( ! is.null(na.action) ) x <- na.action(x)
+  
+  # if( ! is.null(na.action) ) x <- na.action(x)
+  if( na.rm ) x <- na.omit(x)
   
   d <- diff(x)
   if( is.null(step) ) 
@@ -40,8 +43,8 @@ is_incremental <- function(x, step=1, na.action=NULL  ) {
 }
 
 
-#' @rdname is_incremental
+#' @rdname tests
 #' @export 
 
-is_uniform <- function(...) 
-  is_incremental( step=NULL )
+is_uniform <-  function(x, step=1, na.rm=TRUE )
+  is_incremental( x ,step=step, na.rm=na.rm )