+++ NEWS.md
@@ -1,3 +1,10 @@
+# monotone 0.4.1
+
+ - Add is_strictly_sorted as alias for is_strictly_monotonic 
+ - Update DESCRIPTION and README.md
+ - Fix docs
+ - Add tests
+ 
 # monotone 0.4.0 
 
  - Break apart code into smaller files